requirements.txt
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
entrypoints==0.3
flake8==3.7.7
idna==2.8
mccabe==0.6.1
prettytable==0.7.2
prompt-toolkit==2.0.9
pycodestyle==2.5.0
pycryptodome==3.8.0
pyflakes==2.1.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
wcwidth==0.1.7
yapf==0.26.0
